Item detail connected with backend

+++ src/screens/ItemDetailScreen.js
@@ -1,11 +1,13 @@
 import React, { useEffect, useState } from 'react';
-import { View, Text, StyleSheet, Image, TouchableOpacity, ScrollView, Modal, TouchableWithoutFeedback, TextInput } from 'react-native';
+import { View, Text, StyleSheet, Image, TouchableOpacity, ScrollView, Modal, TouchableWithoutFeedback, TextInput, Alert } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import Header from '../components/layout/Header'
 import Footer from '../components/layout/Footer'
-import { productsData } from '../services/products';
 import CartDrawerModal from '../components/cart/CartDrawerModal';
 import { useNavigation } from '@react-navigation/native';
+import resolveProductImage from '../utils/customfunctions';
+import { ProductService } from '../services/ProductService';
+import FontAwesome from 'react-native-vector-icons/FontAwesome5';
 
 const ItemDetail = ({ route }) => {
   const [itemDetail, setItemDetail] = useState({});
@@ -19,23 +21,51 @@ const ItemDetail = ({ route }) => {
       const existingCartJSON = await AsyncStorage.getItem('cart');
       const existingCart = existingCartJSON ? JSON.parse(existingCartJSON) : [];
       const index = existingCart.findIndex(item => item.id === itemDetail.id);
-
+  
       if (index !== -1) {
-        existingCart[index].quantity += quantity; // Update quantity
+        // Check if adding the new quantity exceeds the stock
+        if (existingCart[index].quantity + quantity > itemDetail.quantityInStock) {
+          Alert.alert(
+            "Quantity Exceeded",
+            "You can't add more than the available stock.",
+            [{ text: "OK", onPress: () => console.log("OK Pressed") }]
+          );
+        } else {
+          existingCart[index].quantity += quantity; // Update quantity
+          await AsyncStorage.setItem('cart', JSON.stringify(existingCart));
+          setCartModalVisible(true); // Show the cart modal
+        }
       } else {
-        existingCart.push({ ...itemDetail, quantity }); // Add new item
+        // For new item, check if the initial quantity exceeds the stock
+        if (quantity > itemDetail.quantityInStock) {
+          Alert.alert(
+            "Quantity Exceeded",
+            "You can't add more than the available stock.",
+            [{ text: "OK", onPress: () => console.log("OK Pressed") }]
+          );
+        } else {
+          existingCart.push({ ...itemDetail, quantity }); // Add new item
+          await AsyncStorage.setItem('cart', JSON.stringify(existingCart));
+          setCartModalVisible(true); // Show the cart modal
+        }
       }
-
-      await AsyncStorage.setItem('cart', JSON.stringify(existingCart));
-      setCartModalVisible(true); // Assuming this controls a confirmation modal
     } catch (error) {
       console.error('Error adding item to cart:', error);
     }
   };
 
-  const onQuantityChange = (newQuantity) => {
+  const onQuantityChange = (newQuantity, totalQuantity) => {
     const num = parseInt(newQuantity, 10);
+   
     if (!isNaN(num) && num > 0) {
+      if (num > totalQuantity) {
+        Alert.alert(
+          "Quantity Exceeded",
+          "You can't add more than the available stock.",
+          [{ text: "OK", onPress: () => console.log("OK Pressed") }]
+        );
+        return;
+      }
       setQuantity(num);
     } else {
       setQuantity(1);
@@ -44,54 +74,66 @@ const ItemDetail = ({ route }) => {
   //find item details
   useEffect(() => {
     const { params } = route;
-    const getItem = productsData.find((p) => {
-      return p.id === params?.id
-    })
-    setItemDetail(getItem);
+    const fetchProductDetail = async (id) => {
+      try {
+        const response = await ProductService.FetchProductDetailById(id)
+        setItemDetail(response)
+        // console.log(itemDetail)
+
+      } catch (error) {
+
+      }
+    }
+    fetchProductDetail(params.id)
+
   }, [route]);
+
+  const productImage = resolveProductImage(itemDetail.name);
+
+
   return (
     <>
       <Header />
       <ScrollView style={styles.container} showsVerticalScrollIndicator={false}>
         <TouchableOpacity style={styles.imageContainer} onPress={() => setModalVisible(true)}>
-          <Image source={itemDetail?.image_url} // Replace with your image path
-            style={styles.image} />
+          <Image source={productImage} style={styles.image} />
         </TouchableOpacity>
         <View style={styles.detailsContainer}>
-          <Text style={styles.title}>{itemDetail?.title}</Text>
-          <Text style={styles.price}>${itemDetail?.price}</Text>
+          <Text style={styles.title}>{itemDetail?.name}</Text>
+          <Text style={styles.price}>{itemDetail.variants ? itemDetail.variants[0].sales_price  : "N/A"} PHP </Text>
+          {parseFloat(itemDetail?.quantityInStock) > 0 ?
+            <>
+              <View style={styles.quantityContainer}>
+                <TouchableOpacity onPress={() => onQuantityChange(quantity - 1, itemDetail.quantityInStock)}>
+                  <FontAwesome name="minus" style={styles.quantityButton} />
+                </TouchableOpacity>
+                <TextInput
+                  style={styles.quantityInput}
+                  keyboardType='numeric'
+                  onChangeText={onQuantityChange}
+                  value={quantity.toString()}
+                />
+                <TouchableOpacity onPress={() => onQuantityChange(quantity + 1 , itemDetail.quantityInStock)}>
+                  <FontAwesome name="plus" style={styles.quantityButton} />
+                </TouchableOpacity>
+                <TouchableOpacity
+                style={styles.addToCartButton} disabled={itemDetail?.quantityInStock <= 0}
+                onPress={() => openCartDrawer()}
+              >
+                <Text style={styles.addToCartButtonText}>Add to cart</Text>
+              </TouchableOpacity>
+              </View>
 
-          <View style={styles.quantityContainer}>
-            <TouchableOpacity onPress={() => onQuantityChange(quantity - 1)}>
-              <Text style={styles.quantityButton}>-</Text>
-            </TouchableOpacity>
-            <TextInput
-              style={styles.quantityInput}
-              keyboardType='numeric'
-              onChangeText={onQuantityChange}
-              value={quantity.toString()}
-            />
-            <TouchableOpacity onPress={() => onQuantityChange(quantity + 1)}>
-              <Text style={styles.quantityButton}>+</Text>
-            </TouchableOpacity>
-          </View>
+               </> : <View style={styles.outOfStockContainer}>
+              <Text style={styles.outOfStockText}>Out of Stock</Text>
+            </View>}
 
-          <TouchableOpacity
-            style={styles.addToCartButton} disabled={itemDetail?.quantity <= 0}
-            onPress={() => openCartDrawer()}
-          >
-            <Text style={styles.addToCartButtonText}>{itemDetail?.quantity > 0 ? "Add to cart" : "Out Of Stock"}</Text>
-          </TouchableOpacity>
+          <Text style={styles.contentItem}>{itemDetail.additional_info}</Text>
 
-          <Text style={styles.consistsTitle}>Consists of</Text>
-          {/* Map your contents here */}
 
-          {itemDetail?.contents?.map((content, index) => (
-            <Text key={index} style={styles.contentItem}>• {content}</Text>
-          ))}
-          {/* ... other items */}
+
         </View>
-        <Modal
+        {/* <Modal
           animationType="fade"
           transparent={true}
           visible={modalVisible}
@@ -111,7 +153,7 @@ const ItemDetail = ({ route }) => {
               </TouchableWithoutFeedback>
             </View>
           </TouchableWithoutFeedback>
-        </Modal>
+        </Modal> */}
       </ScrollView>
       <CartDrawerModal
         visible={cartModalVisible}
@@ -139,101 +181,95 @@ const ItemDetail = ({ route }) => {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: 'white',
+    backgroundColor: '#f0f0f0',  // Light grey background for subtle contrast
   },
   imageContainer: {
-    alignItems: 'center',
+    height: 300,
+    width: '100%',
     marginTop: 5,
   },
   image: {
     width: "100%",
-    height: 300,
-    resizeMode: 'contain',
+    height: "100%",
+    resizeMode: 'cover',
+  },
+  outOfStockContainer: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    marginTop: 5,
+  },
+  outOfStockText: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    color: 'red',
+    textAlign: 'center', // Ensure text is centered within its container
   },
   detailsContainer: {
-    padding: 10,
+    padding: 20,
+    backgroundColor: 'white',  // White background for the details container
+    borderTopLeftRadius: 20,
+    borderTopRightRadius: 20,
+    marginTop: -10,
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: -2,
+    },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 5,
   },
   title: {
-    fontSize: 24,
+    fontSize: 28,
     fontWeight: 'bold',
+    color: '#333',  // Dark grey for better readability
   },
   price: {
-    fontSize: 22,
-    color: 'green',
+    fontSize: 25,
+    color: '#3AA040',  // A shade of green for price
     marginVertical: 10,
   },
   quantityContainer: {
     flexDirection: 'row',
     alignItems: 'center',
-    marginVertical: 10,
   },
   quantityButton: {
     fontSize: 20,
     padding: 10,
-    borderRadius: 15,
-    backgroundColor: '#17588e',
-    color: "white",
+    backgroundColor: '#17588e',  // Consistent blue theme
+    color: 'white',
+    borderRadius: 8,
+    margin:10,  // Circular buttons
   },
   quantityInput: {
-    borderRadius: 15,
+    borderRadius: 8,
     borderWidth: 1,
-    borderColor: 'black',
-    padding: 10,
-    marginHorizontal: 10,
-    width: 40,
+    borderColor: '#ccc',
+    paddingVertical: 8,
+    paddingHorizontal: 15,
     textAlign: 'center',
-    color: "black"
+    width: 100,
+    color: '#333',
   },
   addToCartButton: {
-    backgroundColor: '#17588e',
-    padding: 15,
-    borderRadius: 15,
+    backgroundColor: '#3AA040',  // Vibrant orange button
+    padding: 10,
+    borderRadius: 8,
     alignItems: 'center',
-    marginTop: 10,
-    borderRadius: 15
+    marginTop: 0,
   },
   addToCartButtonText: {
     fontWeight: 'bold',
-    color: "white"
-  },
-  consistsTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginTop: 20,
-    marginBottom: 10,
+    color: 'white',
+    fontSize: 20,
   },
   contentItem: {
     fontSize: 16,
     lineHeight: 24,
-  },
-  centeredView: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0, 0, 0, 0.6)', // Dimmed background
-  },
-  modalView: {
-    width: '80%', // Take up 80% of screen width
-    height: '80%', // Take up 80% of screen height
-    backgroundColor: 'transparent',
-    borderRadius: 20,
-    padding: 10,
-    alignItems: 'center',
-    justifyContent: 'center', // Center children vertically
-    shadowColor: '#ffffff',
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 2,
-    elevation: 1,
-  },
-  modalImage: {
-    width: '100%', // Fill the width of the modalView
-    height: '100%', // Fill the height of the modalView
-    resizeMode: 'contain',
+    color: '#666',  // Lighter text for additional info
   },
 });
+
 
 export default ItemDetail
